decompose welcome component
