+++ src/components/welcome/Welcome.jsx
@@ -1,0 +1,10 @@
+import WelcomeWorld from './WelcomeWorld';
+import WelcomeBallCanvas from './WelcomeBallCanvas';
+export default function Welcome({ setEntered }) {
+  return (
+    <>
+      <WelcomeWorld />
+      <WelcomeBallCanvas setEntered={setEntered} />
+    </>
+  );
+}
